Move gcs, orc tests. Start cleaning ORC tests
